# Optimize `Backend::Flatten#flatten_keys`