--- a/lib/i18n/backend/flatten.rb
+++ b/lib/i18n/backend/flatten.rb
@@ -59,7 +59,7 @@ module I18n
       def flatten_keys(hash, escape, prev_key=nil, &block)
         hash.each_pair do |key, value|
           key = escape_default_separator(key) if escape
-          curr_key = [prev_key, key].compact.join(FLATTEN_SEPARATOR).to_sym
+          curr_key = prev_key ? "#{prev_key}#{FLATTEN_SEPARATOR}#{key}".to_sym : key.to_sym
           yield curr_key, value
           flatten_keys(value, escape, curr_key, &block) if value.is_a?(Hash)
         end